Make messages public
